+++ src/gui/gui-key.c
@@ -2042,6 +2042,7 @@ gui_key_focus_command (const char *key, int context,
 {
     struct t_gui_key *ptr_key;
     int matching, debug, rc;
+    unsigned long value;
     char *command, **commands, **ptr_command;
     const char *str_buffer;
     struct t_hashtable *hashtable;
@@ -2099,9 +2100,9 @@ gui_key_focus_command (const char *key, int context,
         str_buffer = hashtable_get (hashtable, "_buffer");
         if (str_buffer && str_buffer[0])
         {
-            rc = sscanf (str_buffer, "%p", &ptr_buffer);
-            if ((rc == EOF) || (rc == 0))
-                ptr_buffer = gui_current_window->buffer;
+            rc = sscanf (str_buffer, "%lx", &value);
+            if ((rc != EOF) && (rc != 0))
+                ptr_buffer = (struct t_gui_buffer *)value;
         }
         if (!ptr_buffer)
             continue;