+++ src/plugins/plugin-script.c
@@ -405,7 +405,7 @@ plugin_script_str2ptr (struct t_weechat_plugin *weechat_plugin,
                        const char *script_name, const char *function_name,
                        const char *str_pointer)
 {
-    void *pointer;
+    unsigned long value;
     int rc;
     struct t_gui_buffer *ptr_buffer;
 
@@ -415,9 +415,9 @@ plugin_script_str2ptr (struct t_weechat_plugin *weechat_plugin,
     if ((str_pointer[0] != '0') || (str_pointer[1] != 'x'))
         goto invalid;
 
-    rc = sscanf (str_pointer, "%p", &pointer);
+    rc = sscanf (str_pointer + 2, "%lx", &value);
     if ((rc != EOF) && (rc >= 1))
-        return pointer;
+        return (void *)value;
 
 invalid:
     if ((weechat_plugin->debug >= 1) && script_name && function_name)