+++ src/plugins/spell/spell-info.c
@@ -37,6 +37,7 @@ spell_info_info_spell_dict_cb (const void *pointer, void *data,
                                const char *arguments)
 {
     int rc;
+    unsigned long value;
     struct t_gui_buffer *buffer;
     const char *buffer_full_name, *ptr_dict;
 
@@ -53,9 +54,10 @@ spell_info_info_spell_dict_cb (const void *pointer, void *data,
     buffer_full_name = NULL;
     if (strncmp (arguments, "0x", 2) == 0)
     {
-        rc = sscanf (arguments, "%p", &buffer);
-        if ((rc != EOF) && (rc != 0) && buffer)
+        rc = sscanf (arguments, "%lx", &value);
+        if ((rc != EOF) && (rc != 0) && value)
         {
+            buffer = (struct t_gui_buffer *)value;
             if (weechat_hdata_check_pointer (weechat_hdata_get ("buffer"),
                                              NULL, buffer))
             {