+++ CHANGELOG.md
@@ -6,7 +6,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 # WeeChat ChangeLog
 
-## Version 4.7.0 (under dev)
+## Version 4.7.0 (2025-07-19)
 
 ### Changed